🔨 Fix compiler error in release build
The type of `PluginRegistrar`'s `messenger()` method was changed from `flutter::BinaryMessenger*` to `flutter::BinaryMessenger&` in the Flutter 3.19.0 stable release.
This commit updates the code to use the new type.

The return value of `DataReader::ReadUInt16()` was changed from `void` to `uint16_t` in a recent C++/WinRT version.
This commit updates the code to use the new return type.