🔨 Refactor Windows platform name logic
This commit refactors the Windows `flutter_blue_plus_windows_plugin.cpp`
to ensure the `adv_name` and `platform_name` fields are always populated
in the advertisement data, even if the local name is empty.

Previously, these fields were only set if `advertisement.LocalName()`
was not empty. Now, they are always set to the string representation
of `advertisement.LocalName()`.